requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cloudstorage==0.9.0
Flask==1.0.3
flask-paginate==0.5.2
Flask-Script==2.0.6
Flask-WTF==0.14.2
google-api-core==1.12.0
google-auth==1.6.3
google-cloud-core==1.0.2
google-cloud-storage==1.16.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
MarkupSafe==1.1.1
nltk==3.4.3
numpy==1.16.4
pandas==0.24.2
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dateutil==2.8.0
python-igraph==0.7.1.post6
python-magic==0.4.15
pytz==2019.1
requests==2.22.0
rsa==4.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
urllib3==1.25.3
Werkzeug==0.15.4
WTForms==2.2.1
